+++ examples/bitcoin-one-address-balance/src/main.ts
@@ -1,7 +1,0 @@
-import { bootstrap } from '@easylayer/bitcoin-crawler';
-import BalanceModel from './models';
-
-bootstrap({
-  Models: [BalanceModel],
-  rpc: true,
-}).catch((error: Error) => console.error(error));