+++ package/src/__tests__/app.module.test.ts
@@ -1,18 +1,18 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { BitcoinAppModule } from '../app.module';
+import { AppModule } from '../app.module';
 
-describe('BitcoinAppModule', () => {
-  let bitcoinAppModule: BitcoinAppModule;
+describe('AppModule', () => {
+  let appModule: AppModule;
 
   beforeEach(async () => {
     const module: TestingModule = await Test.createTestingModule({
-      imports: [BitcoinAppModule],
+      imports: [AppModule],
     }).compile();
 
-    bitcoinAppModule = module.get<BitcoinAppModule>(BitcoinAppModule);
+    appModule = module.get<AppModule>(AppModule);
   });
 
   it('should be defined', () => {
-    expect(bitcoinAppModule).toBeDefined();
+    expect(appModule).toBeDefined();
   });
 });